RoleDao cleanup
