+++ mdm/common/src/main/java/com/gentics/mesh/core/data/dao/PersistingRoleDao.java
@@ -8,6 +8,7 @@ import static io.netty.handler.codec.http.HttpResponseStatus.BAD_REQUEST;
 import static io.netty.handler.codec.http.HttpResponseStatus.FORBIDDEN;
 import static org.apache.commons.lang3.StringUtils.isEmpty;
 
+import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Stream;
@@ -18,6 +19,7 @@ import com.gentics.mesh.cache.PermissionCache;
 import com.gentics.mesh.context.BulkActionContext;
 import com.gentics.mesh.context.InternalActionContext;
 import com.gentics.mesh.core.data.HibBaseElement;
+import com.gentics.mesh.core.data.group.HibGroup;
 import com.gentics.mesh.core.data.perm.InternalPermission;
 import com.gentics.mesh.core.data.role.HibRole;
 import com.gentics.mesh.core.data.user.HibUser;
@@ -25,9 +27,13 @@ import com.gentics.mesh.core.db.Tx;
 import com.gentics.mesh.core.rest.common.GenericRestResponse;
 import com.gentics.mesh.core.rest.common.PermissionInfo;
 import com.gentics.mesh.core.rest.role.RoleCreateRequest;
+import com.gentics.mesh.core.rest.role.RoleResponse;
+import com.gentics.mesh.core.rest.role.RoleUpdateRequest;
 import com.gentics.mesh.core.result.Result;
 import com.gentics.mesh.core.result.TraversalResult;
 import com.gentics.mesh.event.EventQueueBatch;
+import com.gentics.mesh.parameter.GenericParameters;
+import com.gentics.mesh.parameter.value.FieldsSet;
 
 /**
  * A persisting extension to {@link RoleDao}
@@ -139,4 +145,86 @@ public interface PersistingRoleDao extends RoleDao, PersistingDaoGlobal<HibRole>
 	default void setRolePermissions(HibBaseElement element, InternalActionContext ac, GenericRestResponse model) {
 		model.setRolePerms(getRolePermissions(element, ac, ac.getRolePermissionParameters().getRoleUuid()));
 	}
+
+	@Override
+	default void applyPermissions(HibBaseElement element, EventQueueBatch batch, HibRole role, boolean recursive,
+		Set<InternalPermission> permissionsToGrant,
+		Set<InternalPermission> permissionsToRevoke) {
+		element.applyPermissions(batch, role, recursive, permissionsToGrant, permissionsToRevoke);
+	}
+
+	@Override
+	default HibRole create(String name, HibUser creator, String uuid) {
+		HibRole role = createPersisted(uuid);
+		role.setName(name);
+		role.setCreated(creator);
+		role.generateBucketId();
+		addRole(role);
+		return role;
+	}
+
+	@Override
+	default boolean update(HibRole role, InternalActionContext ac, EventQueueBatch batch) {
+		RoleUpdateRequest requestModel = ac.fromJson(RoleUpdateRequest.class);
+		if (shouldUpdate(requestModel.getName(), role.getName())) {
+			// Check for conflict
+			HibRole roleWithSameName = findByName(requestModel.getName());
+			if (roleWithSameName != null && !roleWithSameName.getUuid().equals(role.getUuid())) {
+				throw conflict(roleWithSameName.getUuid(), requestModel.getName(), "role_conflicting_name");
+			}
+
+			role.setName(requestModel.getName());
+			batch.add(role.onUpdated());
+			return true;
+		}
+		return false;
+	}
+
+	@Override
+	default boolean hasPermission(HibRole role, InternalPermission permission, HibBaseElement vertex) {
+		Set<String> allowedUuids = getRoleUuidsForPerm(vertex, permission);
+		return allowedUuids != null && allowedUuids.contains(role.getUuid());
+	}
+
+	@Override
+	default Set<InternalPermission> getPermissions(HibRole role, HibBaseElement element) {
+		Set<InternalPermission> permissions = new HashSet<>();
+		InternalPermission[] possiblePermissions = element.hasPublishPermissions()
+			? InternalPermission.values()
+			: InternalPermission.basicPermissions();
+
+		for (InternalPermission permission : possiblePermissions) {
+			if (hasPermission(role, permission, element)) {
+				permissions.add(permission);
+			}
+		}
+		return permissions;
+	}
+
+	@Override
+	default RoleResponse transformToRestSync(HibRole role, InternalActionContext ac, int level, String... languageTags) {
+		GenericParameters generic = ac.getGenericParameters();
+		FieldsSet fields = generic.getFields();
+
+		RoleResponse restRole = new RoleResponse();
+
+		if (fields.has("name")) {
+			restRole.setName(role.getName());
+		}
+
+		if (fields.has("groups")) {
+			setGroups(role, ac, restRole);
+		}
+		role.fillCommonRestFields(ac, fields, restRole);
+
+		setRolePermissions(role, ac, restRole);
+		return restRole;
+	}
+
+	private void setGroups(HibRole role, InternalActionContext ac, RoleResponse restRole) {
+		for (HibGroup group : role.getGroups()) {
+			restRole.getGroups().add(group.transformToReference());
+		}
+	}
+
 }
